% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBforecast.R, R/CIBpaintGraph.R
\name{CIBforecast}
\alias{CIBforecast}
\title{CIBforecast}
\usage{
CIBforecast(Transitions)

CIBforecast(Transitions)
}
\arguments{
\item{Transitions}{A stochastic transition matrix (rows are presumed to add to 1). Can be generated by any of the function referenced in \code{\link{CIBTransitionCalculators}}}
}
\value{
A list containing a long term forcast vector (a vector of probabilities of world states), a number representing the entropy of this vector, and a number representing the entropy production.

A list containing a long term forcast vector (a vector of probabilities of world states), a number representing the entropy of this vector, and a number representing the entropy production.
}
\description{
A function for analysis stochastic transition matrices, and outputting the main statistics about them.

Take your CIB transition matrix for a particular transition rule, and produce the corresponding graph.
}
\note{
\code{\link{MakeScoreMatrix}} is probably the function you will need directly after this.

\code{\link{MakeScoreMatrix}} is probably the function you will need directly after this.
}
\examples{
data(ExampleCIBdata)
boltzTrans<-LocalBoltzmann(ExampleCIBdata)
forecast<- CIBforecast(boltzTrans)

data(ExampleCIBdata)
boltzTrans<-LocalBoltzmann(ExampleCIBdata)
forecast<- CIBforecast(boltzTrans)
CIBpaintGraph(boltzTrans)
CIBpaintGraph(boltzTrans,forecast[[1]])

}
\author{
Alastair Jamieson Lane. <aja107@math.ubc.ca>

Alastair Jamieson Lane. <aja107@math.ubc.ca>
}
\keyword{CIB}
