% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBgraph.R
\name{CIBgraph}
\alias{CIBgraph}
\title{CIBpaintGraph}
\usage{
CIBgraph(
  Transitions,
  weight = NULL,
  target = 1,
  cutoffMultiplier = 10^-4,
  layout = layout_with_dh
)
}
\arguments{
\item{Transitions}{A stochastic transition matrix (rows are presumed to add to 1). Can be generated by any of the function referenced in \code{\link{CIBTransitionCalculators}}}

\item{weight}{}
}
\value{
NULL. Function generates a picture, but has not return type.
}
\description{
Take your CIB transition matrix for a particular transition rule, and produce the corresponding graph.
}
\note{
\code{\link{MakeScoreMatrix}} is probably the function you will need directly after this.
}
\examples{
data(ExampleCIBdata)
boltzTrans<-LocalBoltzmann(ExampleCIBdata)
forecast<- CIBforecast(boltzTrans)
CIBpaintGraph(boltzTrans,forecast[[1]])
CIBpaintGraph(boltzTrans,forecast[[1]],2)
CIBpaintGraph(boltzTrans,forecast[[1]],2,0)
CIBpaintGraph(boltzTrans,forecast[[1]],2,10^-4,layout.grid)
}
\author{
Alastair Jamieson Lane. <aja107@math.ubc.ca>
}
\keyword{CIB}
