% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/CIBhighlighter.R
\name{CIBhighlighter}
\alias{CIBhighlighter}
\alias{MakeScoreMatrix}
\title{CIBhighlighter}
\usage{
MakeScoreMatrix(TheList)
}
\arguments{
\item{TheList}{a list containing the CIB matrix, and a "shape" vector. The output of \code{\link{InputCibBanner}} is an appropriate input here.}
}
\value{
A matrix describing the transition probability from each world state (rows) to each other world state (columns).
}
\description{
Given ???? create a vector containing the address of all worldstates satisfying the given criteria.
}
\note{
NOT MADE YET
}
\examples{
CIBobject<- InputCibBanner() ##More example goes here.
}
\author{
Alastair Jamieson Lane. <aja107@math.ubc.ca>
}
\keyword{CIB}

