% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBintroduction.R
\name{CIBintroduction}
\alias{CIBintroduction}
\title{Introduction to CIB}
\description{
StochasticCIB is an R package containing a number of function designed to facilitate Cross-Impact Balance analysis.
For a detailed discussion of Cross Impact Balance please see the references below. 

At its most basic, CIB works as follows:

* Suppose we are studying a system governed by a complex web of interactions. Such a system might an ecology, a country, or even an entire bio-economic system of a planet.

* Suppose are able to track numerous "descriptors" of this system. If studying a country we might have such descriptors as "Government", "Economic status" "Population density" and "Diplomatic standing"

* Each descriptor can be in a variety of states. For example the Government might be a military dictatorship, a democracy or a monarchry.

* Suppose we have reason to believe that the status of each descriptor is liable to have an impact on the status of other descriptors: Educational level will affect income. Population density will affect environmental conditions. We record these beliefs numerically in a Cross Impact Matrix

* CIB analysis takes these beliefs and then uses them to give each possible World State a "score" from the point of view of each other world state.

* World states that give themselves a high score are termed "Consistant", and are expected to be stable. World states that give OTHER states are highsccore are unstable, and are liable to "transition" to other states via some form of transition rule.

* Based on these relative scores, and a transition rule, we can "predicted" where a system may end up after some time, and which states are in some sense "paradoxical", and unlikely to be seen.

Please be warned, Cross Impact Balance is nothing more than a mathematical framework for exploring hypothetical scenarios.
It is not a crystal ball, and it would be unwise to imagine that the presence of numbers correspond with either certainty or precision. The numbers merely allow us to mechanise the process of considering hypothetical situation, and allow us to estimate where we should focus our attention.
}
\note{
To get started use \code{\link{InputCibBanner}} to describe the system you are studying.
 Next \code{\link{MakeScoreMatrix}} will find the relative scores of each 
 world state from each other world state, and you'll be able to use one of
  the many \code{\link{TransitionCalculators}} to determine the transition probabilities of your system.
  Finally, \code{\link{CIBforecast}} will calculate the long term forecast.
  
  To see an Example CIB matrix, use \code{data(\link{antCIB})}.
}
\references{
Markov Models of Social Change (Part 1)- Azimuth Forum https://johncarlosbaez.wordpress.com/2014/02/24/markov-models-of-social-change-part-1/

Markov Models of Social Change (Part 2)- Azimuth Forum https://johncarlosbaez.wordpress.com/2014/03/05/markov-models-of-social-change-part-2/

Wolfgang Weimer-Jehle, Cross-impact balances: a system-theoretical approach to cross-impact analysis, Technological Forecasting & Social Change 73 (2006), 334–361.
}
\author{
Alastair Jamieson-Lane. <aja107@math.ubc.ca>
}
