% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleCIBdata.R
\name{ExampleCIBdata}
\alias{ExampleCIBdata}
\title{ExampleCIBdata}
\description{
A simple example CIB matrix for use in the examples.
ExampleCIBdata considers the descriptors Population, Economy and Education. 
Each state can take the value either Low, Medium or High. 
For the sake of this example we use the impact matrix previously considered by 
Lloyd and Schweizer (2014).
For a slightly larger example, please see \code{\link{SomewhereData}}.
}
\examples{
data(ExampleCIBdata)
CIBmatrix<-ExampleCIBdata[[1]]

}
\author{
Alastair Jamieson Lane. <aja107@math.ubc.ca>
}
\keyword{,}
\keyword{CIB}
\keyword{data}
