% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InputCibBanner.R
\name{InputCibBanner}
\alias{InputCibBanner}
\title{InputCibBanner}
\usage{
InputCibBanner()
}
\value{
A list containing first a CIB matrix, and second a vector allowing the computer to more easily interpret said matrix.
}
\description{
A Zero-input function that will prompt the user to list their descriptors, along with the states associated with each descriptors. Useful for getting started with CIB.
}
\note{
\code{\link{MakeScoreMatrix}} is probably the function you will need directly after this.
}
\examples{
CIBobject<- InputCibBanner()
TransRelAdj<-MakeScoreMatrix(CIBobject)
Transitions<-LocalBoltzmann(CIBobject,TransRelAdj)
CIBforecast(Transitions)

}
\author{
Alastair Jamieson Lane. <aja107@math.ubc.ca>
}
\keyword{CIB}
