% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/LocalArctan.R
\name{LocalArctan}
\alias{LocalArctan}
\title{LocalArctan}
\usage{
LocalArctan(TheList, TransRelAdj = NA, beta = 1, shift = 0)
}
\arguments{
\item{TheList}{a list containing the CIB matrix, and a "shape" vector. The output of \code{\link{InputCibBanner}} is an appropriate input here.}

\item{TransRelAdj}{a list containing a blank transition matrix, a relative score matrix, and an adjacency matrix. The output of \code{\link{MakeScoreMatrix}} is appropriate here.}

\item{beta}{Optional parameter used to scale score values. Large beta leads to a sharp `switch' from low to high probabilitiy}

\item{shift}{Optional parameter used to determine location of switch point between low and high probability. By default, this value is 0, indicating that the switch point occurs as a worldstates relative score passes the score of the current worldstate.}
}
\value{
A matrix describing the transition probability from each world state (rows) to each other world state (columns).
}
\description{
A Local stochastic succession rule. For all adjacent states (states where only one descriptor has changed), calculate the relative change in score (compared to staying in the current location). Take the arctan function of this value to get the relative probability. <NOTE, need better way of explaining this, or reference to paper>. Generally speaking, this transition function will give all transitions with negitive score negligable probability, and all transitions with positive score roughly equal probability.
}
\note{
Behaves very similat to \code{\link{LocalLogistic}}
}
\examples{
data(antCIB)
LocalArctan(antCIB)
}
\author{
Alastair Jamieson Lane. <aja107@math.ubc.ca>
}
\keyword{CIB}

