% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LocalLogistic.R
\name{LocalLogistic}
\alias{LocalLogistic}
\title{LocalLogistic}
\usage{
LocalLogistic(TheList, TransRelAdj = NA, beta = 1, shift = 0)
}
\arguments{
\item{TheList}{a list containing the CIB matrix, and a "shape" vector. The output of \code{\link{InputCibBanner}} is an appropriate input here.}

\item{TransRelAdj}{a list containing a blank transition matrix, a relative score matrix, and an adjacency matrix. The output of \code{\link{MakeScoreMatrix}} is appropriate here.}

\item{beta}{Optional parameter used to scale score values. Large beta leads to a sharp `switch' from low to high probabilitiy}

\item{shift}{Optional parameter used to determine location of switch point between low and high probability. By default, this value is 0, indicating that the switch point occurs as a worldstates relative score passes the score of the current worldstate.}
}
\value{
A matrix describing the transition probability from each world state (rows) to each other world state (columns).
}
\description{
A Local stochastic succession rule. For all adjacent states (states where only one descriptor has changed), calculate the relative change in score (compared to staying in the current location). Take the logistic function (1/(exp^-x+1) ) of this value to get the relative probability. 
Stated mathematically, if \eqn{P_{A\rightarrow B}} is the transition rate from A to B, and \eqn{S_A(B)} is the score of the state B, from the point of view of A, then
\deqn{P_{A\rightarrow B= \frac{1/(1+exp[-\beta (S_A(B)-S_A(A))])}}{\sum 1/(1+exp[-\beta (S_A(C)-S_A(A))])} }
Where here we sum over all states that involve changing at most one descriptor.
}
\details{
The net result of this behavior is to make "Negative" transitions exceedingly unlikely, while all ``positive'' transitions recieve roughly equal weight, regardless of score.
}
\note{
Behaves very similat to \code{\link{LocalArctan}}
}
\examples{
data(ExampleCIBdata)
LocalLogistic(ExampleCIBdata)
}
\author{
Alastair Jamieson Lane. <aja107@math.ubc.ca>
}
\keyword{CIB}
