% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SomewhereData.R
\name{SomewhereData}
\alias{SomewhereData}
\title{SomewhereData}
\description{
A slightly larger example CIB matrix, as described by Wolfgang Weimer-Jehle 
in ScenarioWizard manual 4.3.
For a slightly smaller example, please see \code{\link{ExampleCIBdata}}.
}
\examples{
data(SomewhereData)
CIBmatrix<-SomewhereData[[1]]

}
\author{
Alastair Jamieson Lane. <aja107@math.ubc.ca>
}
\keyword{,}
\keyword{CIB}
\keyword{data}
