% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/TransToFastestActing.R
\name{TransToFastestActing}
\alias{TransToFastestActing}
\title{TransToFastestActing}
\usage{
TransToFastestActing(TheList, TransRelAdj = NA)
}
\arguments{
\item{TheList}{a list containing the CIB matrix, and a "shape" vector. The output of \code{\link{InputCibBanner}} is an appropriate input here.}

\item{TransRelAdj}{a list containing a blank transition matrix, a relative score matrix, and an adjacency matrix. The output of \code{\link{MakeScoreMatrix}} is appropriate here. If left blank, can be calculated on the fly.}
}
\value{
A matrix describing the transition probability from each world state (rows) to each other world state (columns).
}
\description{
This transition function assumes that descriptors are written in order of "speed of update" (slowest to fastest). This function looks to see if there is any way to improve score by changing the fastest updating descriptor. If not, then it checks the 2nd fastest, 3rd fastest, and so on.
}
\note{
In the case of a tie, transition probability is split evenly between the top .
}
\examples{
data(antCIB)
TransToFastestActing(antCIB)
}
\author{
Alastair Jamieson Lane. <aja107@math.ubc.ca>
}
\keyword{CIB}

