% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antCIB.R
\name{antCIB}
\alias{antCIB}
\title{antCIB}
\description{
A simple example CIB matrix for use in the examples.
AntCIB mearuse the Population, Economy and Education levels inside a small ant farm, where each state can take the value either Low, Medium or High.
}
\examples{
data(antCIB)
CIBmatrix<-antCIB[[1]]

}
\author{
Alastair Jamieson Lane. <aja107@math.ubc.ca>
}
\keyword{,}
\keyword{CIB}
\keyword{data}
