% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/CIBforecast.R
\name{CIBforecast}
\alias{CIBforecast}
\title{CIBforecast}
\usage{
CIBforecast(Transitions)
}
\arguments{
\item{Transitions}{A stochastic transition matrix (rows are presumed to add to 1). Can be generated by any of the function referenced in \code{\link{CIBTransitionCalculators}}}
}
\value{
A list containing a long term forcast vector (a vector of probabilities of world states), a number representing the entropy of this vector, and a number representing the entropy production.
}
\description{
A function for analysis stochastic transition matrices, and outputting the main statistics about them.
}
\note{
\code{\link{MakeScoreMatrix}} is probably the function you will need directly after this.
}
\examples{
data(antCIB)
}
\author{
Alastair Jamieson Lane. <aja107@math.ubc.ca>
}
\references{
I need a reference for entropy and entropy production.
}
\keyword{CIB}

