% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/MakeScoreMatrix.R
\name{MakeScoreMatrix}
\alias{MakeScoreMatrix}
\title{MakeScoreMatrix}
\usage{
MakeScoreMatrix(TheList)
}
\arguments{
\item{TheList}{A list, the first element of which is a CIB matrix, and the second of which is a vector describing the "shape" of the CIB matrix (a vector listing how many states per descriptor). The output of  \code{\link{InputCibBanner}} is a suitable input for this function.}
}
\value{
A list containing a blank transition matrix, a "Score" matrix, and an adjacency matrix. The score matrix gives the consistancy score of every world state from the point of view of every other (rows being the source state, columns the target state). The adjacency matrix identifies WHICH descriptor would need to be changed to move from one state to another, taking the value 0 if no descriptor is changed, or -1 if multiple descriptors must be changed.
}
\description{
Given a CIB matrix, calculate the consistancy scores of each world state, as viewed from each other world state. Also calculates adjacentcy of states, along with making a empty transition matrix to be filled in later.
}
\note{
The output of this function makes suitable input for various \code{\link{CIBTransitionCalculators}}.
}
\examples{
data(antCIB)
TransRelAdj<-MakeScoreMatrix(antCIB)
Scores<-TransRelAdj[[2]]
Adjacency<-TransRelAdj[[3]]
}
\author{
Alastair Jamieson Lane. <aja107@math.ubc.ca>
}
\keyword{CIB}

