% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransToMax.R
\name{TransToMax}
\alias{TransToMax}
\title{TransToMax}
\usage{
TransToMax(TheList, TransRelAdj = NA)
}
\arguments{
\item{TheList}{a list containing the CIB matrix, and a "shape" vector. The output of \code{\link{InputCibBanner}} is an appropriate input here.}

\item{TransRelAdj}{a list containing a blank transition matrix, a relative score matrix, and an adjacency matrix. The output of \code{\link{MakeScoreMatrix}} is appropriate here.}
}
\value{
A matrix describing the transition probability from each world state (rows) to each other world state (columns).
}
\description{
A transition matrix function, whereby given the current worldstate, we always transition directly to the world state with the highest possible score.
}
\note{
In the case of a tie, transition probability is split evenly between the top .

Other transition functions can be found via \code{\link{CIBTransitionCalculators}}
}
\examples{
data(antCIB)
Transitions<-TransToMax(antCIB)


}
\author{
Alastair Jamieson Lane. <aja107@math.ubc.ca>
}
\keyword{CIB}
